refactor: 엔드포인트 디렉토리 결합
