+++ src/api/endpoints/thread.ts
@@ -1,6 +1,6 @@
 import axios from "axios";
-import { ThreadsResponse } from "../types/thread";
-import axiosClient from "./clients/axiosClient";
+import { ThreadsResponse } from "../../types/thread";
+import axiosClient from "../clients/axiosClient";
 
 interface GetThreadsParams {
   page: number;
